requirements.txt
cryptography==43.0.1
pycryptodome==3.20.0
